Move Hazard xarray reader into separate module
